Updated the docker_memcache repo

+++ scripts/memcache_system_upgrade/1_remove_containers.sh
@@ -8,12 +8,19 @@ declare -A containers
 
 while IFS== read -r key value; do
     containers[$key]=${value}
-    S=$(docker stop $key);
-    R=$(docker rm $key);
-    echo "Stopped:" $S
-    echo "Removed:" $R
-    printf "\n"
-done < "/opt/docker_memcache/files/containers.txt"
+    CS=$(docker ps -a |grep $key);
+    if [[ "$CS" != "" ]]; then
+        S=$(docker stop $key);
+        R=$(docker rm $key);
+        echo "Stopped:" $S
+        echo "Removed:" $R
+        printf "\n"
+    else
+        echo -e "Container $key ${RED}does not exist${NC}"
+        exit 2
+    fi
+printf "\n"
+done < "/opt/docker_memcache/scripts/memcached_system_upgrade/containers.txt"
 printf "\n"
 echo -e "${GREEN}************************************************************************************************************"
 echo -e "${NC}Please!!!, Now you can ${RED}MANUALLY ${NC}perform required upgrades e.g., OS, Kernell, Docker Engine etc. "
